+++ apps/web/server/src/routes/index.js
@@ -2,11 +2,15 @@ const sp_route = require('./spellcheck');
 const public_route = require('./public');
 const auth_route = require('./auth');
 const profile_route = require('./profile');
+const admin_route = require ('./admin');
+const wabap_route = require('./wabap')
 
 function apiRoutes({ app, logger }) {
   app.use(sp_route({ logger }));
   app.use(auth_route({ logger }));
   app.use(profile_route({ logger }));
+  app.use(wabap_route({ logger }));
+  app.use(admin_route({logger}));
 }
 
 function zakariRoutes({ app, logger }) {