+++ libs/web/ui/src/lib/TopNavbar.js
@@ -2,6 +2,7 @@ import { useCallback, useState, useEffect } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
 import { Button, Navbar } from 'react-bulma-components'
 import { useZakari } from '@kreyolopal/react-zakari'
+import SearchBox from '../Wabap/SearchBox'
 const classNames = require('classnames')
 
 const useToggle = (initialState = false) => {
@@ -46,6 +47,9 @@ export const TopNavbar = () => {
       <Navbar.Burger onClick={openMobileMenu} aria-label="menu" />
       <Navbar.Menu renderAs="div" className={navMenu}>
         <Navbar.Container align="right">
+          <Navbar.Item>
+            <SearchBox />
+          </Navbar.Item>
           {isLoggedIn ? (
             <>
               <Navbar.Item href="/spellcheck">Korije</Navbar.Item>